requirements.txt
fastapi>=0.109.0
pandas>=2.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
pytest>=7.4.4
git+https://github.com/haven-jeon/PyKoSpacing.git
konlpy>=0.6.0
networkx>=3.2.1
scipy>=1.12.0
scikit-learn>=1.4.0
transformers>=4.37.2
torch>=2.2.0
urlextract>=1.8.0
emoji>=2.10.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiofiles>=23.2.1
psutil>=5.9.8
numpy>=1.24.0
motor>=3.3.2